[KFI]refactor(pickers-react): moved listpicker props to new file
