+++ packages/sn-pickers-react/src/ListPicker/index.tsx
@@ -1,2 +1,3 @@
 export * from './ListPicker'
 export * from './Item'
+export * from './ListPickerProps'
